feat:用户端取消订单、再来一单&管理端订单搜索、取消订单

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -41,5 +41,5 @@ public interface OrderMapper {
      * @return
      */
     @Select("select * from orders where id = #{id}")
-    Orders getById(Integer id);
+    Orders getById(Long id);
 }
